Cor 6107 add return for rid

+++ src/Laravel/config/config.php
@@ -12,7 +12,7 @@ return [
 	'logger_name'          => env('LOGGING_LOGGER_NAME'),
 	'logger_queue'         => env('LOGGING_LOGGER_QUEUE'),
 	'logger_enable_queue'  => env('LOGGING_LOGGER_ENABLE_QUEUE'),
-	'logger_min_log_level' => env('LOGGING_LOGGER_MIN_LOG_LEVEL', Logger::INFO),
+	'logger_min_log_level' => (int) env('LOGGING_LOGGER_MIN_LOG_LEVEL', Logger::INFO),
 	'graylog_host'         => env('LOGGING_GRAYLOG_HOST'),
-	'graylog_port'         => env('LOGGING_GRAYLOG_PORT'),
+	'graylog_port'         => (int) env('LOGGING_GRAYLOG_PORT'),
 ];
